🧪 test: 스토리 파일명을 변경하고, 컴포넌트에 해당하는 스토리 작성
